Move refcell lint into shared module
